Completed Exercise 4.9
Implemented a "for" loop that accepts an initial value, predicate that
indicates whether the loop should continue, and a step function and
then executes the body of the function with i bound into the context
of the body execution. This is accomplished by setting the initial
value of i and the lambda function to be executed for each iteration
in a let form. Then an inner define is created for a function named
"iter-loop". This function takes the parameters i (the current
iteration variable value) and v (the last execution of f's value) and
if the predicate passes calls iter-loop with an updated binding using
the step function and the execution of to bind to the value of v. If
the predicate fails, the previous value of the execution of f is returned.